### 2022-03-06 #4227 @editorialbot
[PRE REVIEW]: DuoDIC: 3D Digital Image Correlation in MATLAB
**Submitting author:** @danasolav (<a href="http://orcid.org/0000-0003-3215-0401">Dana Solav</a>)
**Repository:** https://github.com/SolavLab/DuoDIC
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
