### 2022-03-16 #4227 @danasolav
Thanks @kthyng. I understand the situation, and am willing to wait. However, please note that this software is not intended for biomechanics. It is related to experimental mechanics and characterization of the mechanical properties of materials and structures. Biomechanics is just one of the fields it can be useful for, but not the only or main one. I hope this information helps in finding another editor soon.
