### 2022-03-28 #4279 @diehlpk
> ```
> Software report:
>
> github.com/AlDanial/cloc v 1.88  T=1.33 s (112.4 files/s, 36797.9 lines/s)
> -------------------------------------------------------------------------------
> Language                     files          blank        comment           code
> -------------------------------------------------------------------------------
> MATLAB                         126           4641          12062          24788
> C++                             15            818           1408           4351
> Markdown                         4            131              0            315
> TeX                              1             16              0            211
> C/C++ Header                     3             68             84            168
> Java                             1              8              7             23
> -------------------------------------------------------------------------------
> SUM:                           150           5682          13561          29856
> -------------------------------------------------------------------------------
>
>
> gitinspector failed to run statistical information for the repository
> ```

@openjournals/dev Could you please have a look?
