### 2022-03-31 #4227 @iagolemos1
Hello,

Sorry for the late answer. I won't be able to review the paper because of
paralel things that are going on, otherwise, I would be happy to help!

Best regards!

Le jeu. 31 mars 2022 à 15:50, Babak N. Safa ***@***.***> a
écrit :

> Hi @iagolemos1 <https://github.com/iagolemos1> or @bleyerj
> <https://github.com/bleyerj> or @BabakNSafa
> <https://github.com/BabakNSafa> would you be interested to review this
> paper?
>
> Sorry for the late reply. Unfortunately, I won't be able to review at this
> point because of multiple pressing deadlines. Thanks for reaching out.
>
