### 2022-05-16 #4279 @abhishektha
Hi @danasolav, great work on this software. I think there are a few things to fix before the software is ready for publication.

I found a few minor errors in the installation document:
- 2.5 Installation:
  - I believe the correct command is install_DuoDIC
  - The last sentence of the installation is repeated twice
- 3.1.2: Missing reference
- 4.2.1.o: Missing reference

Other suggestions from the checklist:
- Could you make it clearer that you can just reload the calibrationSession.mat variable into the workspace to reload calibration parameters for multiple analyses? I think this will help newer Matlab users.
- I think it would be really helpful to point out the sample data in the installation instructions so that users can verify the software is installed and performing correctly. For this, it would be helpful to have the specific parameters used in the example data sets so that the results can be verified. Could you add these to the installation documentation so that I and future users can verify the software is working properly? This includes:
  - checkerboard size used in the sample data
  - DIC parameters

I should be able to verify that the software works once I have these parameters to run through the steps myself, and then compare against the results in the sample data!
