### 2022-05-16 70af859
update instruction manual
Following paper review suggestions by abhishektha
