### 2022-05-16 #4279 @danasolav
Dear @abhishektha ,

Thank you for your helpful suggestions!

I fixed all the minor errors you pointed out.

Regarding the other suggestions:

1. This was mentioned in the last paragraph in 4.1.1. Perhaps it wasn't entirely clear, I tried to rephrase it a bit. I'm not sure what you mean by "reload the variable into the workspace". This can be done but it's not required because in step 3 (3D reconstruction) the user is asked to select the calibrationSession.mat file, and then the selected file is loaded to the workspace. I wanted to clarify that the same file can be used for multiple analysis as long as the images were captured with the same camera settings.
2. I added references to the sample data in the instructions, in Section 3, and along the steps instructions (e.g., 4.1.1c, 4.2.1d, 4.2.1f). The checkerboard square size is 10 mm. You're correct that it wasn't specifically noted, it was only written in the folder name (calibration_11_18_10mm). I have added this detail as a txt file, as well as in the instruction. Regarding the DIC parameters, there is no one "correct" set of parameters to use, and the user can experiment with different settings, but I added the ones I used to generate the figures, for reference.

Please let me know if you have any further questions.
Thanks,
Dana
