### 2022-05-16 #4279 @mr53563
We have built an experimental setup to directly test the code. This took us a little bit since we’ve never used 3D DIC before. We had a little bit of trouble that @danasolav <https://github.com/danasolav> is helping us figuring out right now. Rest assured, we are working hard on getting this done and doing a thorough job. Thank you for checking in!
_____________________________________
Manuel K. Rausch, PhD, Assistant Professor
The University of Texas at Austin | Aerospace Engineering & Engineering Mechanics | cell: 650 283 0262 | www.manuelrausch.com <http://www.manuelrausch.com/>

> On May 16, 2022, at 9:47 AM, Patrick Diehl ***@***.***> wrote:
>
>
> Hi @mr53563 <https://github.com/mr53563> how is your review going?
>
