### 2022-05-18 #4279 @mr53563
Hi @danasolav, I am confused about how references work in the paper. Work is references via @NameYear where are they supposed to be listed? At the bottom of the document, the reference list is empty. Sorry, I am new to this :).
