### 2022-05-18 #4279 @mr53563
@danasolav Also, in the paper, can you include and discuss the limitations of ncorr that you have mentioned in response to our troubles with the number of images and with the large deformation?
