### 2022-05-19 #4279 @danasolav
Hi @mr53563 , I assume you refer to the paper.md file? This file is formatted to a pdf together with the .bib file that contains the references. You can download the [formatted paper](https://github.com/openjournals/joss-papers/blob/joss.04279/joss.04279/10.21105.joss.04279.pdf) to check that all the references appear adequately.
