### 2022-05-24 7dedaf4
update readme and instructions
added limitations: number of images memory issue, and seed placement with large deformation and/or large stereo angle.
