### 2022-05-24 #4279 @danasolav
I have added to the instruction manual and the paper paragraphs describing the two limitations that were raised by @mr53563:
1. Memory issues limiting the number/size of images that can be processed at once.
2. Ncorr's seed placement algorithm struggling to place the seeds correctly in case of large deformations, especially when accompanied by a large stereo angle between the cameras.
