### 2022-05-28 #4279 @mr53563
@danasolav After reducing the image size and the total image number we were able to run the code on our example data set. It took ~24 hours on a desktop machine and in the end we got some strange outputs with clear distortion. Question is now: is it worth exploring this with you to make sure we didn't do anything fundamentally wrong? Or is our data set just junk? Would it make sense for us to provide our data and you try it on your end?
