### 2022-05-30 #4279 @danasolav
@mr53563 I tried processing the data you sent me and I think that the main limiting factor is the large stereo angle between the cameras. This large distortion in the images from the second camera causes Ncorr to struggle with placing the seed, which in turn makes the analysis much longer than it should be. I'm not sure if you noticed, but when seed placement succeeds, the processing time of each image is much shorter than when it fails (because it searches in the entire image instead of finding it right next to where the initial guess was). For this reason, I would not take these running times as an example. In any case, when seed placement fails, the results are meaningless and there is no point in proceeding to the processing step.
If you want, I can help you improve the setup for 3D-DIC analysis, or you could test the sample data, which also contains large strains values (up to ~30%, but without cracks).
