### 2022-06-15 #4279 @diehlpk
@danasolav

I would like to make some editorial comments:

- [ ]  I think it would be beneficial for the reader to mention some closed software products here.

> Commercial 3D-DIC software are typically expensive, proprietary and closed-source, which
> may pose a barrier, especially for students and researchers.

- [ ]  Could you please add a link to the example data?

>
> The figures in this section represent sample data, which are
> provided with the toolbox.

After these changes, I will accept the paper.
