+++ paper/paper.md
@@ -32,23 +32,23 @@ Three-dimensional Digital Image Correlation (3D-DIC) is a non-contact optical-nu
 
 The simple user interface allows novice users to perform 3D-DIC analyses without interacting with MATLAB syntax, while stand-alone functions can be integrated in custom scripts by more proficient MATLAB users. As such, `DuoDIC` is suitable for students, researchers, and professionals in various fields.
 
-The package is composed of four main scripts: (1) stereo camera calibration; (2) image cross-correlation (2D-DIC); (3) 3D reconstruction; and (4) post-processing. This paper describes the algorithms implemented in each step and demonstrates its performance in two test cases, which are also included as sample data: rigid body translations of a cylindrical container and uniaxial tension of a rubber dog-bone specimen.
+The package is composed of four main scripts: (1) stereo camera calibration; (2) image cross-correlation (2D-DIC); (3) 3D reconstruction; and (4) post-processing. This paper describes the algorithms implemented in each step and demonstrates its performance in two test cases, which are also included as [sample data](https://github.com/SolavLab/DuoDIC/tree/main/sample_data): rigid body translations of a cylindrical container and uniaxial tension of a rubber dog-bone specimen.
 
 
 # Statement of need
 
 3D-DIC is useful in numerous applications where the full-field displacements and strains are required, primarily for measuring the mechanical response of materials and structures and for characterizing mechanical properties, with widespread applications from nano- to macro-scale, and rapid progress and developments [@Reu2015; @Sutton2017; @Pan2018]. 3D-DIC is preferable over strain gages and extensometers due to its non-contact nature and the large amount of local information that can be collected.
 
-Commercial 3D-DIC software are typically expensive, proprietary and closed-source, which may pose a barrier, especially for students and researchers. Notable free and open-source 3D-DIC contributions include `DICe` [@Turner2015], `MultiDIC` [@Solav2018], and `ADIC3D` [@Atkinson2021]. Each of these packages uses different algorithms for correlation and stereo calibration and requires different calibration targets. Specifically, our previous toolbox `MultiDIC` focused on multi-view applications, therefore required a non-planar calibration object, which is relatively difficult to make accurate, thus preventing easy implementation. To this end, `DuoDIC` enables a simpler calibration procedure that requires only a flat checkerboard pattern. In addition, with respect to `MultiDIC` we improved the post-processing and visualization options. `DuoDIC` is written in MATLAB, providing the flexibility and simple implementation of a high-level language, which meets the needs of the experimental mechanics community. Moreover, a main advantages of open-source software is the ability to compare results and performance, and to cross-validate the implementation of different algorithms and approaches. To this end, `DuoDIC` may also comprise a complementary tool for researchers in the field.
[…]
