### 2022-06-16 #4279 @danasolav
@diehlpk
I addressed your comments in my last commit.
I wasn't sure if you wanted the link to the sample data to appear in the text or just as a hyperlink (I did the latter). Please let me know if that's ok.
Thanks!
