### 2022-06-19 #4279 @danasolav
@diehlpk
This is the tagged release: https://github.com/SolavLab/DuoDIC/releases/tag/v.1.1.0
And the DOI of the Zenodo archive: https://doi.org/10.5281/zenodo.6664889
Thanks!
