### 2022-06-20 #4279 @diehlpk
@danasolav the on Zenodo reads as

> SolavLab/DuoDIC: DuoDIC: 3D Digital Image Correlation in MATLAB (Release v1.1.0)

and the paper's title reads as

> DuoDIC: 3D Digital Image Correlation in MATLAB

Can you please remove the version number at the end and 'SolavLab/DuoDIC' at the beginning?
