### 2022-06-20 #4279 @danasolav
@diehlpk I updated the Zenodo title to: "DuoDIC: 3D Digital Image Correlation in MATLAB"
