### 2022-06-20 #4279 @diehlpk
> :warning: [Error prepararing paper acceptance](https://github.com/openjournals/joss-papers/actions/runs/2529275755?check_suite_focus=true). The generated XML metadata file is invalid.
>
> ```yaml
> Element isbn: [facet 'minLength'] The value has a length of '6'; this underruns the allowed minimum length of '10'.
> Element isbn: [facet 'pattern'] The value '-98-71' is not accepted by the pattern '(97(8|9)-)?\d[\d \-]+[\dX]'.
> Element isbn: [facet 'minLength'] The value has a length of '9'; this underruns the allowed minimum length of '10'.
> ```
>

@arfon could you please have a look?
