### 2022-06-20 #4279 @xuanxu
There are [some](https://github.com/SolavLab/DuoDIC/blob/main/paper/paper.bib#L61) [invalid](https://github.com/SolavLab/DuoDIC/blob/main/paper/paper.bib#L31) ISBN values in the bib file
