### 2022-06-20 #4279 @diehlpk
> There are [some](https://github.com/SolavLab/DuoDIC/blob/main/paper/paper.bib#L61) [invalid](https://github.com/SolavLab/DuoDIC/blob/main/paper/paper.bib#L31) ISBN values in the bib file

@danasolav could you please check that?
