+++ paper/paper.bib
@@ -28,8 +28,6 @@ arxivId = {arXiv:1011.1669v3},
 author = {Zhang, Zhengyou},
 doi = {10.1109/34.888718},
 eprint = {arXiv:1011.1669v3},
-isbn = {MSR-TR-98-71},
-issn = {MSR-TR-98-71},
 journal = {IEEE Transactions on Pattern Analysis and Machine Intelligence},
 number = {11},
 pages = {1330--1334},
@@ -44,8 +42,6 @@ year = {2000}
 abstract = {In geometrical camera calibration the objective is to determine a\nset of camera parameters that describe the mapping between 3-D reference\ncoordinates and 2-D image coordinates. Various methods for camera\ncalibration can be found from the literature. However surprisingly\nlittle attention has been paid to the whole calibration procedure, i.e.,\ncontrol point extraction from images, model fitting, image correction,\nand errors originating in these stages. The main interest has been in\nmodel fitting, although the other stages are also important. In this\npaper we present a four-step calibration procedure that is an extension\nto the two-step method. There is an additional step to compensate for\ndistortion caused by circular features, and a step for correcting the\ndistorted image coordinates. The image correction is performed with an\nempirical inverse model that accurately compensates for radial and\ntangential distortions. Finally, a linear method for solving the\nparameters of the inverse model is presented},
 author = {Heikkila, J. and Silven, O.},
 doi = {10.1109/CVPR.1997.609468},
-isbn = {0-8186-7822-4},
-issn = {0-8186-7822-4},
 journal = {Proceedings of IEEE Computer Society Conference on Computer Vision and Pattern Recognition},
 pages = {1106--1112},
 pmid = {1151469},
@@ -58,8 +54,6 @@ year = {1997}
 abstract = {Digital Image Correlation (DIC) is an important and widely used non-contact technique for measuring material deformation. Considerable progress has been made in recent decades in both developing new experimental DIC techniques and in enhancing the performance of the relevant computational algorithms. Despite this progress, there is a distinct lack of a freely available, high-quality, flexible DIC software. This paper documents a new DIC software package Ncorr that is meant to fill that crucial gap. Ncorr is an open-source subset based 2D DIC package that amalgamates modern DIC algorithms proposed in the literature with additional enhancements. Several applications of Ncorr that both validate it and showcase its capabilities are discussed.},
 author = {Blaber, J. and Adair, B. and Antoniou, A.},
 doi = {10.1007/s11340-015-0009-1},
-isbn = {0014-4851},
-issn = {0014-4851},
 journal = {Experimental Mechanics},
 keywords = {Complex ROI,Crack,Digital image correlation,Large deformation,Nickel superalloy},
 number = {6},
@@ -104,8 +98,6 @@ year = {2014}
 @article{Solav2017b,
 author = {Solav, Dana and Meric, Henri and Rubin, M. B. and Pradon, Didier and Lofaso, Fr{\'{e}}d{\'{e}}ric and Wolf, Alon},
 doi = {10.1007/s10439-017-1840-6},
-isbn = {1043901718406},
-issn = {15739686},
 journal = {Annals of Biomedical Engineering},
 keywords = {Breathing pattern,Chest wall kinematics,Neuromuscular disorder,Optoelectronic plethysmography,Respiratory motion},
 number = {8},
@@ -164,7 +156,6 @@ year = {2013}
 @book{Bradski2008,
 author = {Bradski, Gary and Kaehler, Adrian},
 edition = {First Edit},
-isbn = {978-0-596-51613-0},
 publisher = {O'Reilly Media, Inc.},
 title = {{Learning OpenCV}},
 url = {http://files/1329/Bradski, Kaehler - 2008 - Learning OpenCV.pdf},
