### 2022-06-21 #4279 @danasolav
@diehlpk I deleted the unnecessary ISBNs. Could you please try again?
The updated release: https://github.com/SolavLab/DuoDIC/releases/tag/v1.1.0
The new DOI: https://doi.org/10.5281/zenodo.6673060
